Se agregan todos los elementos y enemigos.

+++ modelo/NumeroJugador.cpp
@@ -16,5 +16,5 @@ void NumeroJugador::serializar(ostream &stream) {
 }
 
 void NumeroJugador::deserializar(istream &stream) {
-    numeroJugador = static_cast<int>(deserializarEntero(stream));
+    numeroJugador = deserializarEntero(stream);
 }
